produce only a ref-map as it contains all data somebody would want. (#450)
It does more work as well, but it seems unnecessary to maintian a bigger
API surface and it might be key to turning this into a step-by-step
transition to fetching packs and updating local references.
