+++ git-repository/src/remote/connection/mod.rs
@@ -66,4 +66,4 @@ mod access {
 }
 
 ///
-pub mod list_refs;
+pub mod ref_map;
