+++ git-repository/src/remote/mod.rs
@@ -73,7 +73,7 @@ pub mod connect;
 #[cfg(any(feature = "async-network-client", feature = "blocking-network-client"))]
 mod connection;
 #[cfg(any(feature = "async-network-client", feature = "blocking-network-client"))]
-pub use connection::{list_refs, Connection};
+pub use connection::{ref_map, Connection};
 
 mod access;
 pub(crate) mod url;
